feat(DCMAW-8965): rename generateKeyPair file
