+++ sts-mock/jwks-helper-script/publish_keys_to_s3.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-PRIVATE_KEY_FILE_NAME="private-key.json";
-JWKS_FILE_NAME="jwks.json";
+PRIVATE_KEY_JWK_FILE_NAME="private-key.json";
+PUBLIC_KEY_JWKS_FILE_NAME="jwks.json";
 
 aws sts get-caller-identity >/dev/null 2>&1
 if [ $? -gt 0 ]; then
@@ -16,12 +16,12 @@ fi
 
 echo "Generating keys"
 npm install
-node jwksHelperScript.mjs $PRIVATE_KEY_FILE_NAME $JWKS_FILE_NAME
+node generateKeyPair.mjs $PRIVATE_KEY_JWK_FILE_NAME $PUBLIC_KEY_JWKS_FILE_NAME
 
 STACK_NAME="$1"
 ENVIRONMENT="$2"
 BUCKET_NAME="$STACK_NAME"-jwks-"$ENVIRONMENT"
 
 echo "Uploading keys"
-aws s3 cp $PRIVATE_KEY_FILE_NAME s3://"$BUCKET_NAME"/
-aws s3 cp $JWKS_FILE_NAME s3://"$BUCKET_NAME"/.well-known/
+aws s3 cp $PRIVATE_KEY_JWK_FILE_NAME s3://"$BUCKET_NAME"/
+aws s3 cp $PUBLIC_KEY_JWKS_FILE_NAME s3://"$BUCKET_NAME"/.well-known/
